requirements.txt
faster-whisper
flask
flask-cors
flask-jwt-extended
python-dotenv
